easy/code/backend/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
cryptography==36.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.1.6
django-allauth==0.49.0
django-cors-headers==3.7.0
django-countries==7.3.1
django-crispy-forms==1.11.0
django-filter==21.1
django-heroku==0.3.1
django-ip-geolocation==1.6.1
django-rest-auth==0.9.5
djangorestframework==3.12.4
djangorestframework-filter==0.2.2
djangorestframework-simplejwt==5.0.0
gunicorn==20.0.4
idna==2.10
oauthlib==3.2.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.21
PyJWT==2.3.0
python-dotenv==0.19.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.1
typing_extensions==4.1.1
urllib3==1.26.4
whitenoise==5.2.0